+++ src/app/code-block/code-block.component.ts
@@ -21,29 +21,14 @@ import {
         </button>
 
         <button type="button" class="btn-actions code-block-actions btn-menu"
-          soho-context-menu trigger="click">
+          soho-context-menu trigger="click" [beforeOpen]="onBeforeContextMenuOpen"
+          (selected)="onSelected($event)">
           <svg class="icon" focusable="false" aria-hidden="true" role="presentation">
             <use xlink:href="#icon-more"></use>
           </svg>
           <span>Launch</span>
         </button>
-        <ul soho-popupmenu>
-          <li soho-popupmenu-item><a soho-popupmenu-label>Show Field History</a></li>
-          <li soho-popupmenu-item><a soho-popupmenu-label>Show Pending Changes</a></li>
-          <li soho-popupmenu-item>
-            <a soho-popupmenu-label>Drill Around</a>
-            <ul soho-popupmenu>
-              <li soho-popupmenu-item><a soho-popupmenu-label>Sub Menu 1</a></li>
-              <li soho-popupmenu-item><a soho-popupmenu-label>Sub Menu 2</a></li>
-            </ul>
-          </li>
-          <li soho-popupmenu-separator singleSelectableSection="true"></li>
-          <li soho-popupmenu-heading>Options</li>
-          <li soho-popupmenu-item
-          (click)="toggleLabels()" [isChecked]="!hideLabels" isSelectable="true">
-              <a soho-popupmenu-label>Show Labels</a>
-          </li>
-        </ul>
+        <ul class="popupmenu"></ul>
       </div>
     </div>`,
   styleUrls: ['./code-block.component.css']
@@ -61,6 +46,23 @@ export class CodeBlockComponent implements OnDestroy {
 
   isReadOnly = false;
 
+  private MENU_RESPONSE_HTML = '' +
+    '<li><a href="#" id="ShowFieldHistory">Show Field History</a></li>' +
+    '<li><a href="#" id="ShowPendingChanges">Show Pending Changes</a></li>' +
+    '<li class="submenu">' +
+    '<a href="#">Drill Points</a>' +
+    '<ul class="popupmenu"></ul>' +
+    '</li>' +
+    '<li class="separator"></li>' +
+    '<li class="heading">Options</li>' +
+    '<li class="is-selectable"><a href="#" id="ShowLabels">Show Labels</a></li>' +
+    '';
+
+  private SUBMENU_RESPONSE_HTML = '' +
+    '<li><a href="#" id="DrillOne">Drill Point One</a></li>' +
+    '<li><a href="#" id="DrillTwo">Drill Point Two</a></li>' +
+    '';
+
   constructor(private elementRef: ElementRef) {
   }
 
@@ -74,5 +76,24 @@ export class CodeBlockComponent implements OnDestroy {
 
   toggleLabels() {
     setTimeout(() => this.hideLabels = !this.hideLabels);
+  }
+
+  onBeforeContextMenuOpen = (response: AjaxBeforeOpenResponseFunction, options: any) => {
+    if (options.hasOwnProperty('contextElement')) {
+      response(this.SUBMENU_RESPONSE_HTML);
+      return;
+    } else {
+      response(this.MENU_RESPONSE_HTML);
+      return;
+    }
+  }
+
+  onSelected($event) {
+    const buttonId = $event.args[0].id;
+    if (buttonId === 'ShowLabels') {
+      this.toggleLabels();
+    } else {
+      alert(buttonId + ' clicked');
+    }
   }
 }