+++ src/soho/lookup/soho-lookup.component.ts
@@ -55,6 +55,10 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
     this._options.field = value;
   }
 
+  @Input() set match(match: SohoDataGridMatchFunction) {
+    this._options.match = match;
+  }
+
   // Make sure you bind the context to the function
   @Input() set click(value: SohoLookupClickFunction) {
     this._options.click = value;