+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## Whats New
 
 * 2017-12-11 - CWF - SOHO-6856: Upgraded JQuery types from 2.0.46 to 3.2.16 to match our use of jQuery 3
+* 2017-12-06 - MHH - SOHO-7228: Added support for lazy loading the context menu's popup menu.
+* 2017-12-06 - MAF - SOHO-7216: Added support for SohoAlert.
 * 2017-11-07 - BTH - SOHO-6797/SOHO-7015/SOHO-6263: Added support for component formatters and editors for datagrid.
 * 2017-11-17 - BTH - PR-512 - Added 'emptyMessage' support.
 * 2017-11-10 - BTH - PR-503 - Upgraded to angular-cli 1.5 and angular 5.